Cleanup. All tests passing.
